Migrate interventions-list to LitElement
